// fitepic-web-app/src/app/core/api/generated/models/auth-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Authentication response returned after a successful sign-in or registration.
 */
export interface AuthResponse {

  /**
   * The authenticated user's email address.
   */
  email?: string | null;

  /**
   * UTC date and time when the token expires. Obtain a new token via SignIn after this time.
   */
  expiresAt?: string;

  /**
   * The authenticated user's display name. Null if not set at registration.
   */
  name?: string | null;

  /**
   * JWT bearer token. Supply this in the `Authorization: Bearer {token}` header for all authenticated requests.
   */
  token?: string | null;
}